pyproject.toml
[tool.poetry]
name = "easylife"
version = "0.1.0"
description = ""
authors = ["krch-sudo <krv93200@proton.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.2"
celery = "^5.4.0"
psycopg2-binary = "^2.9.9"
dj-database-url = "^2.2.0"
sentry-sdk = "^2.16.0"
django-ninja = "^1.3.0"
pyjwt = "^2.9.0"
django-ninja-jwt = "^5.3.4"
kink = "^0.8.0"
deal = "^4.24.4"


[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"
pytest-django = "^4.9.0"
pytest-env = "^1.1.5"
pytest-mock = "^3.14.0"
pytest-cov = "^5.0.0"


[tool.poetry.group.dev.dependencies]
astroid = "^3.3.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
